Updated upload_image example
The old upload_image() example used fake data and the documentation does not make it readily apparent that you can simply pass the result of open() directly to image_attrs as it is an iterable data source. The original example seems to imply that the data must all be in memory, which is obviously not reasonable for large images.